30. A Better Way to Organize Code
